services/resume-builder-agent/requirements.txt
# Resume Builder Agent Requirements

# Core Framework
langgraph>=0.1.0
pydantic>=2.0
pydantic-settings>=2.0
asyncio>=3.4.3

# PDF Processing
pdfplumber>=0.9.0
reportlab>=4.0.0
pypdf>=3.18.0

# Document Generation
python-docx>=0.8.11

# LLM Integration
anthropic>=0.20.0
openai>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web Requests
aiohttp>=3.8.0
requests>=2.31.0

# AWS Integration
boto3>=1.26.0
botocore>=1.29.0

# Logging
python-json-logger>=2.0.0
python-dotenv>=1.0.0

# Utilities
pyyaml>=6.0
typing-extensions>=4.8.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0
ruff>=0.1.0
